+++ src/main/java/com/github/wintersteve25/tau/components/animated/AnimatedTexture.java
@@ -1,34 +1,32 @@
-package com.github.wintersteve25.tau.components.utils;
+package com.github.wintersteve25.tau.components.animated;
 
 import com.github.wintersteve25.tau.build.BuildContext;
-import com.github.wintersteve25.tau.components.base.UIComponent;
-import com.github.wintersteve25.tau.theme.Theme;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.components.events.GuiEventListener;
-import net.minecraft.client.gui.components.Renderable;
-import net.minecraft.resources.ResourceLocation;
-import com.github.wintersteve25.tau.components.base.DynamicUIComponent;
 import com.github.wintersteve25.tau.components.base.PrimitiveUIComponent;
+import com.github.wintersteve25.tau.components.base.UIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
+import com.github.wintersteve25.tau.theme.Theme;
 import com.github.wintersteve25.tau.utils.SimpleVec2i;
+import com.github.wintersteve25.tau.utils.Size;
+import net.minecraft.resources.ResourceLocation;
 
-import javax.transaction.xa.Xid;
-import java.util.List;
+import java.util.function.Supplier;
 
-public final class Texture implements PrimitiveUIComponent {
+public class AnimatedTexture implements PrimitiveUIComponent {
 
     private final ResourceLocation textureLocation;
     private final SimpleVec2i textureSize;
-    private final SimpleVec2i uv;
     private final SimpleVec2i uvSize;
     private final SimpleVec2i size;
+    private final Supplier<Size> uvSizeAnimator;
+    private final Supplier<SimpleVec2i> uvPosAnimator;
 
-    public Texture(ResourceLocation textureLocation, SimpleVec2i textureSize, SimpleVec2i uv, SimpleVec2i uvSize, SimpleVec2i size) {
+    public AnimatedTexture(ResourceLocation textureLocation, SimpleVec2i textureSize, SimpleVec2i uvSize, SimpleVec2i size, Supplier<Size> uvSizeAnimator, Supplier<SimpleVec2i> uvPosAnimator) {
         this.textureLocation = textureLocation;
         this.textureSize = textureSize;
-        this.uv = uv;
         this.uvSize = uvSize;
         this.size = size;
+        this.uvSizeAnimator = uvSizeAnimator;
+        this.uvPosAnimator = uvPosAnimator;
     }
 
     @Override
@@ -37,7 +35,11 @@ public final class Texture implements PrimitiveUIComponent {
         SimpleVec2i position = layout.getPosition(uvSize);
 
         context.renderables().add((graphics, pMouseX, pMouseY, pPartialTicks) -> {
-            graphics.blit(textureLocation, position.x, position.y, size.x, size.y, uv.x, uv.y, uvSize.x, uvSize.y,
+            SimpleVec2i newUvSize = new SimpleVec2i(uvSize.x, uvSize.y);
+            newUvSize = uvSizeAnimator.get().get(newUvSize);
+            SimpleVec2i newUvPos = uvPosAnimator.get();
+
+            graphics.blit(textureLocation, position.x, position.y, size.x, size.y, newUvPos.x, newUvPos.y, newUvSize.x, newUvSize.y,
                     textureSize.x,
                     textureSize.y);
         });
@@ -48,22 +50,20 @@ public final class Texture implements PrimitiveUIComponent {
     public static final class Builder implements UIComponent {
 
         private final ResourceLocation textureLocation;
+        private final Supplier<Size> uvSizeAnimator;
+        private final Supplier<SimpleVec2i> uvPosAnimator;
         private SimpleVec2i textureSize;
-        private SimpleVec2i uv;
         private SimpleVec2i uvSize;
         private SimpleVec2i size;
 
-        public Builder(ResourceLocation textureLocation) {
+        public Builder(ResourceLocation textureLocation, Supplier<Size> uvSizeAnimator, Supplier<SimpleVec2i> uvPosAnimator) {
             this.textureLocation = textureLocation;
+            this.uvSizeAnimator = uvSizeAnimator;
+            this.uvPosAnimator = uvPosAnimator;
         }
 
         public Builder withTextureSize(SimpleVec2i textureSize) {
             this.textureSize = textureSize;
-            return this;
-        }
-
-        public Builder withUv(SimpleVec2i uv) {
-            this.uv = uv;
             return this;
         }
 
@@ -77,10 +77,10 @@ public final class Texture implements PrimitiveUIComponent {
             return this;
         }
 
-        public Texture build() {
+        public UIComponent build() {
             textureSize = textureSize == null ? new SimpleVec2i(256, 256) : textureSize;
             uvSize = uvSize == null ? textureSize : uvSize;
-            return new Texture(textureLocation, textureSize, uv == null ? SimpleVec2i.zero() : uv, uvSize, size == null ? uvSize : size);
+            return new AnimatedTexture(textureLocation, textureSize, uvSize, size == null ? uvSize : size, uvSizeAnimator, uvPosAnimator);
         }
 
         @Override