+++ src/main/java/com/github/wintersteve25/tau/menu/TauContainerMenu.java
@@ -1,7 +1,7 @@
 package com.github.wintersteve25.tau.menu;
 
+import com.github.wintersteve25.tau.Tau;
 import net.minecraft.core.BlockPos;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.*;
@@ -10,49 +10,60 @@ import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class TauContainerMenu extends AbstractContainerMenu implements ContainerListener {
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import java.util.function.Supplier;
 
-    public final CompoundTag data;
+public class TauContainerMenu extends AbstractContainerMenu {
+
     public final Level level;
     public final BlockPos pos;
 
-    public TauContainerMenu(@Nullable MenuType<?> pMenuType, Inventory playerInventory, int pContainerId, BlockPos pos) {
-        super(pMenuType, pContainerId);
+    private final Map<String, IndexedDataSlot> dataSlots;
+    private final TauMenuHolder holder;
 
-        this.data = new CompoundTag();
+    public TauContainerMenu(TauMenuHolder holder, Inventory playerInventory, int pContainerId, BlockPos pos) {
+        super(holder.get(), pContainerId);
+
         this.level = playerInventory.player.level();
         this.pos = pos;
-
-        addSlotListener(this);
+        this.dataSlots = new HashMap<>();
+        this.holder = holder;
     }
 
     @Override
     public ItemStack quickMoveStack(Player pPlayer, int pIndex) {
-        return null;
+        return holder.getMenu().quickMoveStack(this, pPlayer, pIndex);
     }
 
     @Override
     public boolean stillValid(Player pPlayer) {
-        return true;
-    }
-
-    @Override
-    public void slotChanged(AbstractContainerMenu pContainerToSend, int pDataSlotIndex, ItemStack pStack) {
-    }
-
-    @Override
-    public void dataChanged(AbstractContainerMenu pContainerMenu, int pDataSlotIndex, int pValue) {
+        return holder.getMenu().stillValid(this, pPlayer);
     }
 
     public @NotNull Slot addSlot(@NotNull Slot pSlot) {
         return super.addSlot(pSlot);
     }
 
-    public @NotNull DataSlot addDataSlot(@NotNull DataSlot pIntValue) {
-        return super.addDataSlot(pIntValue);
+    public void addDataSlot(String name, DataSlot slot) {
+        if (dataSlots.containsKey(name)) {
+            Tau.LOGGER.error("Duplicated data slot key: " + name);
+            return;
+        }
+
+        dataSlots.put(name, new IndexedDataSlot(dataSlots.size(), slot));
+        addDataSlot(slot);
     }
 
-    public void addDataSlots(@NotNull ContainerData pArray) {
-        super.addDataSlots(pArray);
+    public Optional<Supplier<Integer>> getGetterForDataSlot(String name) {
+        if (!dataSlots.containsKey(name)) {
+            Tau.LOGGER.error("Unknown data slot key: " + name);
+            return Optional.empty();
+        }
+
+        return Optional.of(dataSlots.get(name).dataSlot()::get);
     }
+
+    private record IndexedDataSlot(int index, DataSlot dataSlot) {}
 }