+++ CHANGELOG.md
@@ -19,8 +19,7 @@
 
 # v2.0.1 Neoforge-1.20.6 Minor Update
 ## API Changes
-- Added `RenderedUIComponent` as a new component base alongside of `DynamicUIComponent` `PrimitiveUIComponent` and `UIComponent`
-- Added `DataDependentComponent` and `DataDependentRender` for interacting with data slots in a Menu setting
 - Renamed `TauMenu` to `UIMenu`
 - Added `TauContainerMenu containerMenu` param to `UIMenu#build`
 - Added `TauMenuHolder#openMenu`
+- Added `UIMenu#tick`, `UIMenu#addDataSlots` and `UIMenu#stillValid`